Create deamon.py
This is the deamon tool of my clearwell reports.  I intually only had one script, but as this grew I needed to create a client and daemon script that used a network share to overcome security and network hurdles when accessing the internal db and external smtp server.  This is a very early stage and will be much cleaner in the future.  It works though!